refactor: organize benchmarks folder
