+++ benchmarks/runner/benchmarks/from-entry-hash.js
@@ -1,11 +1,12 @@
 const startIPFS = require('./utils/start-ipfs')
 const releaseRepo = require('./utils/release-repo')
 const createLog = require('./utils/create-log')
-const Log = require('../src/log')
+const Log = require('../../../src/log')
 
 const base = {
   prepare: async function () {
-    const { ipfs, repo } = await startIPFS('./ipfs-log-benchmarks/fromMultihash/ipfs')
+    const { ipfs, repo } = await startIPFS('./ipfs-log-benchmarks/fromEntryHash/ipfs')
+
     const { log, access, identity } = await createLog(ipfs, 'A')
     const refCount = 64
     process.stdout.clearLine()
@@ -13,12 +14,18 @@ const base = {
       process.stdout.write(`\r${this.name} / Preparing / Writing: ${i}/${this.count}`)
       await log.append('hello' + i, refCount)
     }
-
-    const multihash = await log.toMultihash()
-    return { ipfs, repo, access, identity, log, multihash }
+    return { ipfs, repo, log, access, identity }
   },
-  cycle: async function ({ log, access, identity, ipfs, multihash }) {
-    await Log.fromMultihash(ipfs, access, identity, multihash)
+  cycle: async function ({ log, ipfs, access, identity }) {
+    await Log.fromEntryHash(
+      ipfs,
+      access,
+      identity,
+      log.heads.map(e => e.hash),
+      log._id,
+      -1,
+      []
+    )
   },
   teardown: async function ({ repo }) {
     await releaseRepo(repo)
@@ -41,8 +48,8 @@ const counts = [1, 100, 1000, 10000]
 let benchmarks = []
 for (const count of counts) {
   const c = { count }
-  if (count < 1000) benchmarks.push({ name: `fromMultihash-${count}-baseline`, ...base, ...c, ...baseline })
-  benchmarks.push({ name: `fromMultihash-${count}-stress`, ...base, ...c, ...stress })
+  if (count < 1000) benchmarks.push({ name: `fromEntryHash-${count}-baseline`, ...base, ...c, ...baseline })
+  benchmarks.push({ name: `fromEntryHash-${count}-stress`, ...base, ...c, ...stress })
 }
 
 module.exports = benchmarks
