+++ benchmarks/runner/benchmarks/has.js
