+++ benchmarks/runner/benchmarks/tail-hashes.js
