+++ benchmarks/runner/benchmarks/to-multihash.js
